text-to-sql/requirements.txt
pyautogen
spider_env